requirements.txt
pytest==8.2.2
selenium==4.22.0
webdriver_manager==4.0.1
allure-pytest==2.13.5
requests==2.32.3